# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/interactions/like:
    post:
      description: Record a positive swipe interaction for a dish type.
      operationId: InteractionsController_likeDish
      parameters: []
      responses:
        '201':
          description: ''
      security: &ref_1
        - access-token: []
      summary: Like a dish
      tags: &ref_2
        - interactions
  /api/v1/interactions/save:
    post:
      description: Add a restaurant to the user's saved collection.
      operationId: InteractionsController_saveRestaurant
      parameters: []
      responses:
        '201':
          description: ''
      security: *ref_1
      summary: Save restaurant
      tags: *ref_2
  /api/v1/interactions/saved:
    get:
      description: Return all restaurants saved by the user.
      operationId: InteractionsController_getSaved
      parameters: []
      responses:
        '200':
          description: ''
      security: *ref_1
      summary: List saved restaurants
      tags: *ref_2
  /api/v1/interactions/saved/{id}:
    delete:
      description: Delete one saved restaurant by interaction id.
      operationId: InteractionsController_deleteSaved
      parameters:
        - name: id
          required: true
          in: path
          description: Saved interaction UUID
          schema:
            format: uuid
      responses:
        '200':
          description: ''
      security: *ref_1
      summary: Remove saved restaurant
      tags: *ref_2
...
info:
  title: Tinner API
  description: >-
    REST API for Tinner — auth, dish suggestions, filters, and saved restaurants. Most routes require JWT Bearer token
    unless marked public.
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Access token from POST /api/v1/auth/login or /register
...